+++ lib/EventProfiler.php
@@ -2,6 +2,10 @@
 
 namespace ICanBoogie;
 
+use ArrayIterator;
+use ICanBoogie\EventProfiler\CallRecord;
+use ICanBoogie\EventProfiler\UnusedRecord;
+
 use function microtime;
 
 /**
@@ -10,40 +14,56 @@ use function microtime;
 final class EventProfiler
 {
     /**
-     * Unused event types.
-     *
-     * ```php
-     * [ $time, $type ] = $value;
-     * ````
-     *
-     * @var array<array{ float, string }>
+     * @var UnusedRecord[]
      */
-    public static array $unused = [];
+    private static array $unused = [];
 
     /**
      * Adds an unused event type.
      */
     public static function add_unused(string $type): void
     {
-        self::$unused[] = [ microtime(true), $type ];
+        self::$unused[] = new UnusedRecord(microtime(true), $type);
+    }
+
+    public static function reset_unused(): void
+    {
+        self::$unused = [];
+    }
+
+    /**
+     * @return iterable<UnusedRecord>
+     */
+    public static function iter_unused(): iterable
+    {
+        return new ArrayIterator(self::$unused);
     }
 
     /**
      * Event hooks calls.
      *
-     * ```php
-     * [ $time, $type, $hook, $started_at ] = $value;
-     * ```
-     *
-     * @var array<array{ float, string, callable, float }>
+     * @var CallRecord[]
      */
-    public static array $calls = [];
+    private static array $calls = [];
 
     /**
      * Adds an event hook call.
      */
     public static function add_call(string $type, callable $hook, float $started_at): void
     {
-        self::$calls[] = [ microtime(true), $type, $hook, $started_at ];
+        self::$calls[] = new CallRecord($started_at, microtime(true) - $started_at, $type, $hook);
+    }
+
+    public static function reset_calls(): void
+    {
+        self::$calls = [];
+    }
+
+    /**
+     * @return iterable<CallRecord>
+     */
+    public static function iter_calls(): iterable
+    {
+        return new ArrayIterator(self::$calls);
     }
 }